scripts/requirements.txt
pandas==2.2.3
DateTime==5.5
requests==2.32.3
python-dotenv==1.0.1
psycopg2-binary==2.9.10